Organized AlbumArt.tsx
